Kernel: Implement device tree overlay with the gpio-line-names description
With the currently employed linux-rpi branch, the gpio line names are not described
by the device tree. To ease the future integration of the custom LKM for gpio lines
interface, the line names should be described. For such, this commit applies a patch
that implements an overlay with the gpio-line-names property defined.

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>