Kernel: Integrate device tree overlay with the gpio-line-names description
With the currently employed linux-rpi branch, the gpio line names are not described
by the device tree. To ease the future integration of the custom LKM for gpio lines
interface, the line names should be described. For such, this commit integrates the
dtbo target into the list of device trees to be deployed in the image and passed to
the boot firmware and kernel through as system configuration parameter.

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>